+++ docs/src/.vitepress/theme/index.ts
@@ -16,6 +16,7 @@ import { enhanceAppWithTabs } from 'vitepress-plugin-tabs/client'
 
 import '@nolebase/vitepress-plugin-enhanced-readabilities/client/style.css'
 import './style.css'
+import './docstrings.css'
 
 export const Theme: ThemeConfig = {
   extends: DefaultTheme,