+++ src/AlgebraOfGraphics.jl
@@ -45,6 +45,8 @@ export scale, scales
 export pregrouped
 export direct
 export from_continuous
+export wrapped
+export clipped
 
 include("dict.jl")
 include("theme.jl")