+++ src/algebra/select.jl
@@ -1,6 +1,41 @@
 struct DimsSelector{N}
     dims::NTuple{N, Int}
 end
+
+"""
+    dims(args...)
+
+Create a `DimsSelector` object which can be used in `mapping` to refer
+to the dimensions of the layer's shape.
+
+Example:
+
+```julia
+mapping([:x1, :x2, :x3], col = dims(1))
+```
+
+In the normal case where only
+single columns are referenced in `mapping`, the shape of the layer is `(1,)`.
+If, for example, the first `mapping` entry, let's say for the X scale,
+is a column vector with three column selectors, like `[:x1, :x2, :x3]`,
+the shape becomes `(3,)`. If there's additionally a second `mapping` entry for
+the Y scale with four column selectors in a row vector, like `[:y1 :y2 :y3 :y4]`,
+which by itself has shape `(1, 4)`, the overall
+shape becomes `(3, 4)`. This means that 3 x 4 = 12 combinations of columns will
+be plotted together as X and Y in the same axis.
+
+The `dims` selector can now be used to create what can be thought of as a categorical
+scale with as many entries as the product of the sizes of the selected dimensions.
+For example, if we have our shape `(3, 4)` and add `color = dims(1)`, the data will
+be colored in three different shades, one for each column selector in the first dimension
+(the three columns passed to the X scale). If we set `color = dims(1, 2)`, there will
+be 12 different shades, one for each combination of X and Y. If we set `color = dims(2)`,
+there will be four different shades, one for each Y column.
+
+With wide input data, it often makes sense to put each column into its own facet, in
+order not to overcrowd the visual space. Therefore, a common setting in the `mapping`
+in such scenarios could be `row = dims(1), col = dims(2)`, for example.
+"""
 dims(args...) = DimsSelector(args)
 
 function (d::DimsSelector)(c::CartesianIndex{N}) where N