+++ src/guides/colorbar.jl
@@ -6,36 +6,60 @@ function colorbar!(fg::FigureGrid; position=:right,
 end
 
 """
-    colorbar!(figpos, grid; kwargs...)
+    colorbar!(figpos, grid; kwargs...)::Vector{Colorbar}
 
-Compute colorbar for `grid` (which should be the output of [`draw!`](@ref)) and draw it in
-position `figpos`. Attributes allowed in `kwargs` are the same as `MakieLayout.Colorbar`.
+Compute zero or more colorbars for `grid` (which should be the output of [`draw!`](@ref)) and draw them in a
+nested `GridLayout` in position `figpos`. One colorbar will be drawn for each applicable scale.
+Attributes allowed in `kwargs` are the same as `MakieLayout.Colorbar`.
+
+!!! note
+    Before AlgebraOfGraphics v0.10, this function returned `Union{Nothing,Colorbar}`
+    because multiple colorbars in a single figure were not supported. The name of the
+    function was kept singular **colorbar** to be less breaking.
 """
-function colorbar!(figpos, grid; kwargs...)
-    colorbar = compute_colorbar(grid)
-    return isnothing(colorbar) ? nothing : Colorbar(figpos; colorbar..., kwargs...)
+function colorbar!(figpos, grid; vertical = true, kwargs...)
+    colorbars = compute_colorbars(grid)
+    return [Colorbar(figpos[(vertical ? (1, i) : (i, 1))...]; colorbar..., vertical, kwargs...) for (i, colorbar) in enumerate(colorbars)]
 end
 
-compute_colorbar(fg::FigureGrid) = compute_colorbar(fg.grid)
-function compute_colorbar(grid::Matrix{AxisEntries})
-    colorscales = filter(!isnothing, [get(ae.continuousscales, AesColor, nothing) for ae in grid])
-    isempty(colorscales) && return
+compute_colorbars(fg::FigureGrid) = compute_colorbars(fg.grid)
 
-    # colorscale = reduce(mergescales, colorscales)
+function should_use_colorbar(colorscale::CategoricalScale)::Bool
+    aesprops::CategoricalAesProps = colorscale.props.aesprops
+    return aesprops.colorbar === Makie.automatic ? is_binned(colorscale) : aesprops.colorbar
+end
+
+function compute_colorbars(grid::Matrix{AxisEntries})
+    colorbars = []
+    catscales = get(first(grid).categoricalscales, AesColor, nothing)
+    if catscales !== nothing
+        for catscale in catscales
+            if should_use_colorbar(catscale)
+                push!(colorbars, categorical_colorbar(catscale))
+            end
+        end
+    end
+    
+    colorscales = filter(!isnothing, [get(ae.continuousscales, AesColor, nothing) for ae in grid])
+
     colorscale_dict = reduce(colorscales, init=Dictionary{Union{Nothing,Symbol},ContinuousScale}()) do c1, c2
         mergewith!(mergescales, c1, c2)
     end
 
-    if length(colorscale_dict) > 1
-        error("Cannot yet handle multiple colorscales, found $colorscale_dict")
+    for colorscale in values(colorscale_dict)
+        push!(colorbars, continuous_colorbar(colorscale))
     end
 
-    colorscale = only(values(colorscale_dict))
-    
+    return colorbars
+end
+
+function continuous_colorbar(colorscale::ContinuousScale)
     label = getlabel(colorscale)
-    limits = @something colorscale.props.aesprops.colorrange colorscale.extrema
+    limits = nonsingular_colorrange(colorscale)
     is_highclipped = limits[2] < colorscale.extrema[2]
     is_lowclipped = limits[1] > colorscale.extrema[1]
+
+    _, limits = strip_units(colorscale, collect(limits))
 
     colormap = @something colorscale.props.aesprops.colormap default_colormap()
     colormap_colors = Makie.to_colormap(colormap)
@@ -50,4 +74,75 @@ function compute_colorbar(grid::Matrix{AxisEntries})
         lowclip,
         highclip,
     )
+end
+
+function is_binned(scale::CategoricalScale)
+    all(x -> x isa Bin, datavalues(scale))
+end
+
+function categorical_colorbar(scale::CategoricalScale)
+    pv = plotvalues(scale)
+    dv = datavalues(scale)
+    label = getlabel(scale)
+
+    if is_binned(scale)
+        bins = sort(dv, by = x -> x.range[2])
+        colors = RGBAf[]
+        values = eltype(bins[1].range)[]
+        for i in eachindex(bins)
+            bin = bins[i]
+            if i == 1
+                push!(values, bin.range...)
+                push!(colors, pv[i])
+            else
+                prev_bin = bins[i-1]
+                # if bins don't touch, we have to insert transparent gaps in between
+                if prev_bin.range[2] > bin.range[1]
+                    error("Overlapping bins in categorical colorscale: $(prev_bin.range) and $(bin.range).")
+                elseif prev_bin.range[2] < bin.range[1]
+                    push!(values, bin.range[1])
+                    push!(colors, RGBAf(0, 0, 0, 0))
+                end
+                push!(values, bin.range[2])
+                push!(colors, pv[i])
+            end
+        end
+
+        if !isfinite(values[1])
+            lowclip = popfirst!(colors)
+            popfirst!(values)
+        else
+            lowclip = nothing
+        end
+
+        if !isfinite(values[end])
+            highclip = pop!(colors)
+            pop!(values)
+        else
+            highclip = nothing
+        end
+
+        limits = extrema(values)
+        cgrad_stops = (values .- limits[1]) ./ (limits[2] - limits[1])
+        colormap = cgrad(colors, cgrad_stops; categorical = true)
+        return (;
+            limits,
+            colormap,
+            label,
+            lowclip,
+            highclip,
+        )
+    else
+        limits = (0.5, length(pv) + 0.5)
+        # reverse so values read top-to-bottom
+        ticks = (1:length(pv), reverse(datalabels(scale)))
+        colormap = cgrad(reverse(pv), range(0, 1, length = length(pv) + 1); categorical = true)
+        return (;
+            limits,
+            colormap,
+            ticks,
+            label,
+        )
+    end
+    
 end